[REG-2000] Move all tests to SDK and fix asmdef
