+++ src/RGUnityBots/Assets/RegressionGames/Resources/RGBotSequences.asset.meta
@@ -1,7 +1,8 @@
 fileFormatVersion: 2
-guid: ab84c266e8dce48e5ac042c249a1b71a
-AssemblyDefinitionImporter:
+guid: abfa8995defb04e4a870aad588a92467
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 11400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
