--- src/RGUnityBots/Assets/Tests.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 8ac5423d121a844d3a2dbfffae3e3aad
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
