--- src/RGUnityBots/Assets/Tests/Editor.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 79cabbf37cda479aadec836588d9c852
-timeCreated: 1702073355
