--- src/RGUnityBots/Assets/Tests/Editor/RegressionGames.Editor.Tests.asmdef
@@ -1,13 +1,0 @@
-{
-    "name": "RegressionGames.Editor.Tests",
-    "optionalUnityReferences": [
-        "TestAssemblies"
-    ],
-    "includePlatforms": [
-        "Editor"
-    ],
-    "references": [
-        "RegressionGames",
-        "RegressionGames.Editor"
-    ]
-}
