--- src/RGUnityBots/Assets/Tests/Editor/RegressionGames.Editor.Tests.asmdef.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 67cf4cd79358b410abe597ad5dda06de
-AssemblyDefinitionImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
