--- src/RGUnityBots/Assets/Tests/Runtime.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 521d0eba22d0427f9bc9df18d07fa24a
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
