--- src/RGUnityBots/Assets/Tests/Runtime/RGUtilsTests.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 4a4f02c2be71452091645d4eae6048ea
-timeCreated: 1702073471
