--- src/RGUnityBots/Assets/Tests/Runtime/RegressionGames.Tests.asmdef
@@ -1,15 +1,0 @@
-{
-    "name": "RegressionGames.Tests",
-    "optionalUnityReferences": [
-        "TestAssemblies"
-    ],
-    "references": [
-        "RegressionGames",
-        "RegressionGames.TestFramework",
-        "Unity.InputSystem",
-        "Unity.InputSystem.TestFramework"
-    ],
-    "precompiledReferences": [
-        "Newtonsoft.Json.dll"
-    ]
-}
