+++ src/gg.regression.unity.bots/Tests/TestFramework/Scripts/LegacyInput.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 89a783eb18e64f3fbf11682b4911904a
+timeCreated: 1726754140
